Fix name of field to predict
In the README file we are told that we will be predicting the `SalesPrice` field. Actually, the correct name of the field in the dataset is `SalePrice`. It might not have led to any confusion but still, better err on the side of safety. Also fixed the name of the dataset file by removing the extra s.